MAR-413 - add users as contributors automatically when a note is edited, created or deleted
